requirements.txt
asgiref==3.8.1
azure-core==1.30.2
azure-storage-blob==12.21.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
idna==3.7
isodate==0.6.1
pycparser==2.22
pymongo==3.12.3
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
